Parse without using String.split
