--- components/resources/library/src/commonMain/kotlin/org/jetbrains/compose/resources/intl/PluralConditionParseException.kt
@@ -1,8 +1,0 @@
-/*
- * Copyright 2020-2024 JetBrains s.r.o. and respective authors and developers.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
- */
-
-package org.jetbrains.compose.resources.intl
-
-internal class PluralConditionParseException(description: String) : Exception("Invalid LDML: $description")
